python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meshlogic"
version = "0.1.0"
description = "Official Python SDK for the MeshLogic API"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    { name = "Mesh Logic Pty Ltd", email = "support@meshlogic.ai" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["meshlogic", "security", "monitoring", "ebpf", "endpoint"]

dependencies = [
    "httpx>=0.24.0",
    "websocket-client>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[project.urls]
Homepage = "https://meshlogic.ai"
Documentation = "https://meshlogic.ai/docs/sdk"
Repository = "https://github.com/MeshLogic-Pty-Ltd/meshlogic-sdk"
Issues = "https://github.com/MeshLogic-Pty-Ltd/meshlogic-sdk/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/meshlogic",
]

[tool.hatch.build.targets.wheel]
packages = ["meshlogic"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py38"]
